implementing python remotecontrol module
