+++ .gitignore
@@ -52,3 +52,4 @@ node_modules
 .kotlin
 
 __pycache__/
+*.egg-info/
