+++ remotecontrol/python/ocpp_emulator_remote_control/__init__.py
@@ -1,0 +1,12 @@
+"""Client library for the OCPP emulator's control socket.
+
+See docs/cli/cli-plan.md (in the ocpp-emulator repo) for the protocol design.
+"""
+
+from .client import ControlClient, DEFAULT_PORT, parse_charge_point_connector
+
+__all__ = [
+    "ControlClient",
+    "DEFAULT_PORT",
+    "parse_charge_point_connector",
+]
