+++ remotecontrol/python/pyproject.toml
@@ -1,0 +1,12 @@
+[build-system]
+requires = ["setuptools>=68"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "ocpp-emulator-remote-control"
+version = "0.1.0"
+description = "Client library for the OCPP emulator's remote-control socket"
+requires-python = ">=3.10"
+
+[tool.setuptools.packages.find]
+include = ["ocpp_emulator_remote_control*"]
